adds component folder for dashboard view
